Simplex-DW/dags/requirements.txt
apache-airflow == 2.0.2
apache-airflow-providers-snowflake == 1.2.0
apache-airflow-providers-postgres == 1.0.2
apache-airflow-providers-slack == 3.0.0
snowflake-connector-python == 2.7.2
fsspec==2022.01.0
httpx==0.21.1
snowflake-connector-python[pandas]
psycopg2-binary==2.8.6
python-dateutil~=2.8.1
SQLAlchemy~=1.3.24
pandas==1.2.4
apache-airflow[postgres_to_s3_operator]
apache-airflow[s3_to_snowflake_operator]
simple_salesforce==1.11.6
Slack-webclient-logger == 2.2
numpy==1.26.2
matplotlib==3.8.2

